% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lib.R
\name{add_lib}
\alias{add_lib}
\title{Add a custom library for matching}
\usage{
add_lib(file, class_name = NULL, and_cols = "all", or_cols = "rest",
  mode = c("Pos", "Neg"), adduct = "[M+H]+", dda = TRUE,
  aif = TRUE)
}
\arguments{
\item{file}{CSV file or a data frame, with lipid names as first column,
and precursor M/Z in second column and a column for each potentail fragment.}

\item{class_name}{Full name of the lipid class. Defaults to basename of the
input \code{file}}

\item{and_cols}{Character or numeric vector for Columns containing all
necessary fragments for ID. #' Defaults to 'all', i.e. all fragments
must be observed in data.}

\item{or_cols}{Character or numeric vector for Columns containing fragments
where at least one fragment must be observed for ID. Defaults to 'rest',
i.e. all fragments not included in \code{and_cols}. Set argument to \code{FALSE} to
have no \code{or_cols}.}

\item{mode}{Whether this library is for 'Pos' or 'Neg' mode. Default is
'Pos'.}

\item{adduct}{Adduct of molecules in the library. Defualt is \code{'[M+H]+'}}

\item{dda}{Whether this library is suitable for DDA data.}

\item{aif}{Whether this library is suitable for AIF data.}
}
\value{
A tibble with a single row containing the library information and
data. The created library is also added to the full list of available
libraries.
}
\description{
Add a custom library for matching
}
\examples{
csv_file <- system.file("extdata", "PC_H.csv", package = "lipID")
lib_data <- readr::read_csv(csv_file)
head(lib_data)

amended_libs <- add_lib(
  csv_file, class_name = 'MyPhosphatidylCholine',
  and_cols = c(2, 12), or_cols = c(8, 9),
  mode = 'Pos', adduct = '[M+H]+',
  dda = TRUE, aif = TRUE
)
amended_libs

# Get only the added library
amended_libs \%>\% dplyr::filter(user_defined)
}
