% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lib.R
\name{add_lib_collection}
\alias{add_lib_collection}
\title{Bulk add multiple libraries for lipid matching}
\usage{
add_lib_collection(rule_file, path = ".")
}
\arguments{
\item{rule_file}{CSV file containing the library list and rules.
Use the template provided in the examples below.}

\item{path}{Folder path containing a list of CSV files for libraries.
All csv files listed in \code{rule_file} must be present. See examples below.}
}
\value{
A tibble with all added libraries. The created libraries
are also added to the full list of available libraries.
}
\description{
Bulk add multiple libraries for lipid matching
}
\examples{
# File cotanining list of libraries and their matching rules
rule_file <- system.file("extdata", "library_info.csv", package = "lipID")
readr::read_csv(rule_file)

# Folder containing csv files listed in rule_file
lib_dir <- system.file("extdata", package = "lipID")

mylibs <- add_lib_collection(rule_file, path = lib_dir)
mylibs

# Get only the added libraries
mylibs \%>\% dplyr::filter(user_defined)
}
