% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipID-package.R
\docType{data}
\name{annotated_features}
\alias{annotated_features}
\title{Example output data with annonated features}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11174 rows and 29 columns.}
\usage{
data("annotated_features")
}
\description{
Example output data with annonated features
}
\examples{
data("annotated_features")
}
\seealso{
lipID
}
\keyword{datasets}
