% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lib.R
\name{get_libs}
\alias{get_libs}
\title{Get libraries for matching with specific mode and type}
\usage{
get_libs(mode = c("Pos", "Neg"), acq = c("dda", "aif"))
}
\arguments{
\item{mode}{Acquisition mode. Default is 'Pos'}

\item{acq}{Acquisition type, either \code{dda} or \code{aif}. Default is 'dda'}
}
\value{
A tibble containing all libraries for the chosen mode and type.
}
\description{
Get libraries for matching with specific mode and type
}
\examples{
get_libs(mode = 'Pos', acq = 'dda')
}
