% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipID-package.R
\docType{data}
\name{librules}
\alias{librules}
\title{A tibble with all available lipid libraries.
Available libraries are based on LipidMatch package. User-defined libraries
can be added with `add_lib` and `add_lib_collection`.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 218 rows and 11 columns.}
\usage{
data("librules")
}
\description{
A tibble with all available lipid libraries.
Available libraries are based on LipidMatch package. User-defined libraries
can be added with `add_lib` and `add_lib_collection`.
}
\examples{
data("librules")
}
\seealso{
add_lib add_lib_collection
}
\keyword{datasets}
