% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipID-package.R
\docType{package}
\name{lipID-package}
\alias{lipID-package}
\title{Fast Lipid Identification from MS/MS spectra}
\description{
\code{lipID} is an extremely fast implementation for rule-based matching of
lipid compounds from MS/MS spctra. While the package logic and libraries
is based on LipidMatch pakcage, the implementation has been completely
rewritten to improve the performance (100x speed), and make it compatible
with the latest R releases and tideverse workflows.\code{lipID} also allows
users to add custom libraries to search against.
}
\author{
Ahmed Mohamed \email{ahmed.mohamed@qimrberghofer.edu.au}
}
