% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{lipID}
\alias{lipID}
\title{Annotate untargeted lipidomics using rule-based matching}
\usage{
lipID(ms2_files, libs, features = NULL, ppm_tol = 30,
  intensity_cutoff = 1000, mz_window = 1, rt_window = 2,
  partial_match_cutoff = 1)
}
\arguments{
\item{ms2_files}{a character vector of file names or directory
containing ms2 files}

\item{libs}{Which libraries to match against. Should be the output of
\code{[get_lib]} or \code{[create_lib]}}

\item{features}{CSV file with the first two columns corresponding to
m/z and RT. Optional.}

\item{ppm_tol}{Mass error tolerance between acquired MS2 fragments and
library. Default tolerance is 30 ppm.}

\item{intensity_cutoff}{Minimum intensity value for MS2 fragments to be
taken into account when matching. Default is 1000}

\item{mz_window}{M/Z window for merging MS2 data with MS1 features.
Ignored if \code{features} is \code{NULL}.}

\item{rt_window}{Retention time window for merging MS2 data with MS1
features. Ignored if \code{features} is \code{NULL}.}

\item{partial_match_cutoff}{Numeric value between 0-1. Allows molecules that
satisfied some, but not all of the rules to be retained. Default is \code{1},
returning molecules with 100\% match. Set to \code{0} to include molecules
that are MS1 matched as well (i.e no rules are net, only the precursor).}
}
\value{
A data frame with these columns:\itemize{
\item ms2_file, precursor, ms2_rt   File, precursor M/Z, precursor RT
\item name   Name of the matching molecules
\item partial_match   Numeric value between 0-1, indicating the
percentage of rules satisfied. \code{1} indicates matching all
required fragments and at least one optional fragment. \code{0} indicates
the molecule was matched based on MS1 only.
\item confirmed   Whether all matching rules were satisfied.
}
If features table is provided, it is merged with the above columns.
}
\description{
This is a wrapper function that executes the whole workflow
with 2 inputs: \code{ms2_files} and optional \code{features} table.
}
\examples{
ms2_file <- system.file("extdata", "ms2file.ms2", package = "lipID")
features_file <- system.file("extdata", "features.csv", package = "lipID")
libs <- get_libs()
annotated <- lipID(ms2_file, libs, features_file)
head(annotated)
}
