% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1match.R
\name{match_ms1}
\alias{match_ms1}
\title{Match lipid molecules using MS1 precursor mass}
\usage{
match_ms1(features, libs, mz_tol = 0.025)
}
\arguments{
\item{features}{two column dataframe, m/z and RT}

\item{libs}{Which libraries to match agaist. Should be the output of
`[get_lib]` or `[create_lib]`}

\item{mz_tol}{M/Z tolerance for matching, in Da. Default 0.025 Da.}
}
\value{
The input dataframe with additional column
}
\description{
Match lipid molecules using MS1 precursor mass
}
\examples{
features_file <- system.file("extdata", "features.csv", package = "lipID")
features <- readr::read_csv(features_file)
head(features)
libs <- get_libs()

ms1_annotation <- match_ms1(features, libs)
head(ms1_annotation)
}
