% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2match.R
\name{match_ms2}
\alias{match_ms2}
\title{Match lipid molecules from MS2 data}
\usage{
match_ms2(ms2, libs, ppm_tol = 30, intensity_cutoff = 1000,
  collapse = TRUE, odd_chain = FALSE, chain_modifs = c("all", "only",
  "none"))
}
\arguments{
\item{ms2}{MS2 data frame. Should be the result of \code{read_ms2}.}

\item{libs}{Which libraries to match against. Should be the output of
\code{[get_lib]} or \code{[create_lib]}}

\item{ppm_tol}{Mass error tolerance between acquired fragments and
library. Default tolerance is 30 ppm.}

\item{intensity_cutoff}{Minimum intensity value for fragments to be
taken into account when matching. Default is 1000}

\item{collapse}{Whether to collapse ambiguous molecules if they
have the same sum composition. \code{TRUE} by default.}

\item{odd_chain}{Whether to include molecules with odd chain fatty acids.}

\item{chain_modifs}{Whether to include / exclude molecules with modified
chain fatty acids.
\code{FALSE} by default, since odd chains are unlikely in mammals.}
}
\value{
A data frame with these columns:\itemize{
\item ms2_file, precursor, ms2_rt   File, precursor M/Z, precursor RT
\item name   Name of the matching molecules
\item partial_match   Numeric value between 0-1, indicating the
percentage of rules satisfied. \code{1} indicates matching all
required fragments and at least one optional fragment. \code{0} indicates
the molecule was matched based on MS1 only.
\item confirmed   Whether all matching rules were satisfied.
}
This returned values contain all molecules that have a matching
precursor regardless of whether the rules are met or not. If
you want only molecules that met the rules, filter the dataframe
with \code{confirmed == TRUE}.
}
\description{
Rule-based matching of lipid molecules in the provided libraries \code{libs}
using MS2 data (fragments).
}
\examples{
ms2_file <- system.file("extdata", "ms2file.ms2", package = "lipID")

ms2_data <- read_ms2(ms2_file)
libs <- get_libs()
confirmed_molecules <- match_ms2(ms2_data, libs)
head(confirmed_molecules)

# Get only the molecules that satisfied all the rules.
confirmed_molecules \%>\% dplyr::filter(confirmed)
}
