% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2match.R
\name{merge_ms2}
\alias{merge_ms2}
\title{Merge MS2 data with MS1 features}
\usage{
merge_ms2(features, ms2_data, mz_window = 1, rt_window = 2)
}
\arguments{
\item{features}{two column dataframe, m/z and RT}

\item{ms2_data}{MS2 data frame. Should be the result of \code{read_ms2}.}

\item{mz_window}{M/Z window.}

\item{rt_window}{Retention time window.}
}
\value{
A merged data frame containing both MS1 features with their
corresponding MS2 data.
}
\description{
Merge MS2 data with MS1 features
}
\examples{
ms2_file <- system.file("extdata", "ms2file.ms2", package = "lipID")
features_file <- system.file("extdata", "features.csv", package = "lipID")
features <- readr::read_csv(features_file)
head(features)

ms2_data <- read_ms2(ms2_file)
libs <- get_libs()
confirmed_molecules <- match_ms2(ms2_data, libs)
annotated_features <- merge_ms2(features, confirmed_molecules)
head(annotated_features)
}
