% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipID-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
Result of \code{rhs(lhs, ...)}.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
