% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_features}
\alias{plot_features}
\title{Interactively plot identified features}
\usage{
plot_features(results, color_by = c("class_name", "annotated",
  "assigement"))
}
\arguments{
\item{results}{Annotated features from \code{[lipID]} or \code{[merge_ms2]}.}

\item{color_by}{Color points by whether they "annotated",
how they are assigned: "assigement", or their "class_name".}
}
\value{
interactive plot
}
\description{
Interactively plot identified features
}
