% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms2.R
\name{read_ms2}
\alias{read_ms2}
\title{Read MS2 files}
\usage{
read_ms2(files)
}
\arguments{
\item{files}{a character vector of file names or directory
containing ms2 files}
}
\value{
data frame with 6 columns: file name, scan_idx, precursor mz, rt,
fragment mz and fragment intensity.
}
\description{
Read MS2 files
}
\examples{
ms2_file <- system.file("extdata", "ms2file.ms2", package = "lipID")

ms2_data <- read_ms2(ms2_file)
head(ms2_data)
}
