% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1match.R
\name{get_precursors}
\alias{get_precursors}
\title{Lists precusor masses of all molecules in library}
\usage{
get_precursors(libs)
}
\arguments{
\item{libs}{Which libraries to match agaist. Should be the output of
\code{[get_libs]} or \code{[create_lib]}.}
}
\value{
Data frame with three columns: file (name of library),
name (Molecule name) and prec_mz (M/Z of the molecule precusor).
}
\description{
Lists precusor masses of all molecules in library
}
\examples{
libs = get_libs()
precs = get_precursors(libs)
head(precs)
}
